requirements.txt
flask
flask-restful
flask-marshmallow
flask-swagger
flask-login
flask-wtf
flask-httpauth
requests
google-api-python-client